Create 01Lqc_GrabWQ_LongXLinTidyOut.R
This is an alternate version of the file 01_GrabWQ_XLinTidyOut.R for dealing with data submitted with duplicate samples in their own row rather than in a separate column with the same submitted sample metadata.  This Long QC format requires a variable to identify whether a row of data is a regular sample, field primary or field duplicate.  This format is similar to the data that would be downloaded from the Liquid application for field data collection.